Перенёс метод в Presenter
