+++ DavaiPosmotrim/SplashScreen/SplashScreenPresenter.swift
@@ -5,11 +5,15 @@
 //  Created by Maksim Zimens on 05.06.2024.
 //
 
-import Foundation
+import UIKit
+import CoreMotion
 
 final class SplashScreenPresenter: SplashScreenPresenterProtocol {
     weak var coordinator: SplashScreenCoordinator?
     weak var view: SplashScreenViewControllerProtocol?
+
+    private var motion = CMMotionManager()
+    private var queue = OperationQueue.current
 
     init(coordinator: SplashScreenCoordinator) {
         self.coordinator = coordinator
@@ -19,5 +23,17 @@ final class SplashScreenPresenter: SplashScreenPresenterProtocol {
         guard let coordinator else { return }
         coordinator.finish()
     }
-// TODO: Перенести ещё методы из ViewController
+
+    func startMoving(gravity: UIGravityBehavior) {
+        guard let queue = queue else { return }
+        motion.startDeviceMotionUpdates(to: queue) { motion, _ in
+            guard let motion = motion else { return }
+            let grav: CMAcceleration = motion.gravity
+            let xAxis = CGFloat(grav.x)
+            let yAxis = CGFloat(grav.y)
+            let point = CGPoint(x: xAxis, y: yAxis)
+            let vector = CGVector(dx: point.x, dy: 0 - point.y)
+            gravity.gravityDirection = vector
+        }
+    }
 }
