+++ DavaiPosmotrim/SplashScreen/SplashScreenPresenterProtocol.swift
@@ -5,8 +5,9 @@
 //  Created by Maksim Zimens on 05.06.2024.
 //
 
-import Foundation
+import UIKit
 
 protocol SplashScreenPresenterProtocol: AnyObject {
     func splashDidFinish()
+    func startMoving(gravity: UIGravityBehavior)
 }
